requirements.txt
dnspython==1.16.0
mysql-connector-python==8.0.19
numpy==1.18.2
pandas==1.0.3
protobuf==3.6.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
dnspython==1.16.0
mysql-connector-python==8.0.19
numpy==1.18.2
pandas==1.0.3
progress==1.5
protobuf==3.6.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
